requirements.txt
streamlit==1.28.1
pandas==2.1.3
numpy==1.24.3
openpyxl==3.1.2
requests==2.31.0
python-dotenv==1.0.0
msal==1.24.1
plotly==5.17.0
catboost==1.2.2
scikit-learn==1.3.2
joblib==1.3.2
